pyproject.toml
[project]
name = "agent-app"
version = "0.1.0"
description = ""
authors = [
    {name = "gmarcinek",email = "grzegorz.marcinek@gmail.com"}
]
readme = "README.md"
requires-python = "^3.13"
dependencies = [
    "openai (>=1.79.0,<2.0.0)",
    "pydantic (>=2.11.4,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
flake8 = "^7.2.0"

